+++ genium/src/test/resources/smoke/defaults/output/cpp/include/smoke/TypesWithDefaults.h
@@ -36,7 +36,6 @@ struct _GENIUM_CPP_EXPORT ImmutableStructWithDefaults {
     const ::std::string string_field = "\\Jonny \"Magic\" Smith\n";
     const ::smoke::SomeEnum enum_field = ::smoke::SomeEnum::BAR_VALUE;
     const ::fire::SomeVeryExternalEnum external_enum_field = ::fire::SomeVeryExternalEnum::Another_Value;
-    ImmutableStructWithDefaults( );
     ImmutableStructWithDefaults( const uint32_t uint_field, const bool bool_field );
     ImmutableStructWithDefaults( const int32_t int_field, const uint32_t uint_field, const float float_field, const double double_field, const bool bool_field, const ::std::string& string_field, const ::smoke::SomeEnum enum_field, const ::fire::SomeVeryExternalEnum external_enum_field );
 };