+++ genium/src/test/resources/smoke/defaults/output/cpp/src/smoke/TypesWithDefaults.cpp
@@ -8,20 +8,13 @@
 // -------------------------------------------------------------------------------------------------
 #include "smoke/TypesWithDefaults.h"
 namespace smoke {
-StructWithDefaults::StructWithDefaults( )
-    : int_field( 42 ), uint_field( 4294967295 ), float_field( 3.14f ), double_field( -1.4142 ), bool_field( true ), string_field( "\\Jonny \"Magic\" Smith\n" ), enum_field( ::smoke::SomeEnum::BAR_VALUE )
-{
-}
+StructWithDefaults::StructWithDefaults( ) = default;
 StructWithDefaults::StructWithDefaults( const int32_t int_field, const uint32_t uint_field, const float float_field, const double double_field, const bool bool_field, const ::std::string& string_field, const ::smoke::SomeEnum enum_field )
     : int_field( int_field ), uint_field( uint_field ), float_field( float_field ), double_field( double_field ), bool_field( bool_field ), string_field( string_field ), enum_field( enum_field )
 {
 }
-ImmutableStructWithDefaults::ImmutableStructWithDefaults( )
-    : int_field( 42 ), uint_field{ }, float_field( 3.14f ), double_field( -1.4142 ), bool_field{ }, string_field( "\\Jonny \"Magic\" Smith\n" ), enum_field( ::smoke::SomeEnum::BAR_VALUE ), external_enum_field( ::fire::SomeVeryExternalEnum::Another_Value )
-{
-}
 ImmutableStructWithDefaults::ImmutableStructWithDefaults( const uint32_t uint_field, const bool bool_field )
-    : int_field( 42 ), uint_field( uint_field ), float_field( 3.14f ), double_field( -1.4142 ), bool_field( bool_field ), string_field( "\\Jonny \"Magic\" Smith\n" ), enum_field( ::smoke::SomeEnum::BAR_VALUE ), external_enum_field( ::fire::SomeVeryExternalEnum::Another_Value )
+    : uint_field( uint_field ), bool_field( bool_field )
 {
 }
 ImmutableStructWithDefaults::ImmutableStructWithDefaults( const int32_t int_field, const uint32_t uint_field, const float float_field, const double double_field, const bool bool_field, const ::std::string& string_field, const ::smoke::SomeEnum enum_field, const ::fire::SomeVeryExternalEnum external_enum_field )