+++ genium/src/test/resources/smoke/interfaces/output/cpp/src/smoke/InterfaceWithStruct.cpp
@@ -6,20 +6,12 @@
 // Automatically generated. Do not modify. Your changes will be lost.
 //
 // -------------------------------------------------------------------------------------------------
-
 #include "smoke/InterfaceWithStruct.h"
 namespace smoke {
-
 InterfaceWithStruct::~InterfaceWithStruct() = default;
-
-InterfaceWithStruct::InnerStruct::InnerStruct( )
-    : value{ }
-{
-}
-
+InterfaceWithStruct::InnerStruct::InnerStruct( ) = default;
 InterfaceWithStruct::InnerStruct::InnerStruct( const int8_t value )
     : value( value )
 {
 }
-
 }