+++ genium/src/test/resources/smoke/structs_with_companion/output/cpp/src/smoke/StructsWithConstants.cpp
@@ -10,10 +10,7 @@
 namespace smoke {
 const ::std::string Route::DEFAULT_DESCRIPTION = "Nonsense";
 const ::smoke::RouteType Route::DEFAULT_TYPE = ::smoke::RouteType::EQUESTRIAN;
-Route::Route( )
-    : description{ }, type{ }
-{
-}
+Route::Route( ) = default;
 Route::Route( const ::std::string& description, const ::smoke::RouteType type )
     : description( description ), type( type )
 {