+++ genium/src/test/resources/smoke/structs_with_companion/output/cpp/src/smoke/StructsWithConstantsInterface.cpp
@@ -11,10 +11,7 @@ namespace smoke {
 StructsWithConstantsInterface::~StructsWithConstantsInterface() = default;
 const ::std::string StructsWithConstantsInterface::MultiRoute::DEFAULT_DESCRIPTION = "Foo";
 const ::smoke::RouteType StructsWithConstantsInterface::MultiRoute::DEFAULT_TYPE = ::smoke::RouteType::NONE;
-StructsWithConstantsInterface::MultiRoute::MultiRoute( )
-    : descriptions{ }, type{ }
-{
-}
+StructsWithConstantsInterface::MultiRoute::MultiRoute( ) = default;
 StructsWithConstantsInterface::MultiRoute::MultiRoute( const ::std::vector< ::std::string >& descriptions, const ::smoke::RouteType type )
     : descriptions( descriptions ), type( type )
 {